folder renaming map to mapAlgorithms
