+++ _common/gameModel/libs/ClientScript/mapAlgorithms/lineOfSight.ts
