+++ model/gameModel/gamemodel.json
@@ -6466,7 +6466,7 @@
   }, {
     "@class" : "GameModelContent",
     "content" : "",
-    "contentKey" : "map/layersData",
+    "contentKey" : "mapAlgorithms/layersData",
     "contentType" : "application/typescript",
     "libraryType" : "ClientScript",
     "refId" : "GameModelContent:427114:jexCPN",
@@ -6475,7 +6475,7 @@
   }, {
     "@class" : "GameModelContent",
     "content" : "",
-    "contentKey" : "map/lineOfSight",
+    "contentKey" : "mapAlgorithms/lineOfSight",
     "contentType" : "application/typescript",
     "libraryType" : "ClientScript",
     "refId" : "GameModelContent:24073812:LHvCPq",
@@ -6484,7 +6484,7 @@
   }, {
     "@class" : "GameModelContent",
     "content" : "",
-    "contentKey" : "map/pathFinding",
+    "contentKey" : "mapAlgorithms/pathFinding",
     "contentType" : "application/typescript",
     "libraryType" : "ClientScript",
     "refId" : "GameModelContent:441099:z0DmoU",
